Update warning bypass allowlist logic to allowlist by host only
Previously the allowlist required everything except the path to match.
This CL modifies the logic to check the canonicalized host only. The
main goal is to stop checking the scheme, since there can be cases where
specifying http vs. https does not match. But this is expanded to omit
other parts of the URL such as port or user info with similar reasoning.

This CL also adds a histogram logging the scheme on writes to the
allowlist to validate assumptions:
 - SafeBrowsing.WarningBypassAllowlist.SchemeOnWrite

Bug: 356167129
Change-Id: Ia5509c603b7e6dc725fc3000a1a4ce77badd29c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5769742
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339796}