[WebAuthn] Add isCrossOrigin to MakeCredential clientData on Android
On desktop, `isCrossOrigin` is added to the `clientDataJson` for
MakeCredential requests. However, this was never plumbed to the
Android credential creation code, probably because Play Services
generated the clientData and the hash in most cases so it wouldn't
immediately be useful.

This change generates the `clientDataJson` on the Chrome side, and
makes it so that the `clientDataHash` is passed on GMSCore API calls
in all situations. The cross-origin bit is returned from the content
security checks and used.

Change-Id: Ic565663d88323b258ec2d7d103792ca3327e2938
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5774343
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339805}