digital-credentials: add a cross-device transaction class.
This replaces the FidoRequestHandlerBase, which the initial sketch of
this code used. The FidoRequestHandlerBase brings in a _lot_ of other
stuff which this code doesn't need to be entangled with, albeit at the
cost of needing to redo BLE handling.

Change-Id: I2f54501c33d7d9c34498d8be2fe39847f70fb561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5736806
Auto-Submit: Adam Langley <agl@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339813}