[fuchsia] Treat CHROME_HEADLESS as running_unattended
CHROME_HEADLESS is set in both chromium and eureka and can be a
good signal to verify the bot run.
This change partially reverts https://crrev.com/c/4112522 which
looks weird.

But to ensure a soft transition, this change only adds
CHROME_HEADLESS. A TODO is added to remove SWARMING_SERVER once we
force_running_unattended() in dart and flutter.

Bug: 356356430
Change-Id: Icc72eca71f66de22383b018aac64594378083822
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5772679
Commit-Queue: Zijie He <zijiehe@google.com>
Reviewed-by: Guocheng Wei <guochengwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339826}