Fix dangling pointer crash when process is ready after auction ends.
In a previous change -
https://chromium-review.googlesource.com/c/chromium/src/+/5740146 - we
added a metric that measures the start/end time of a phase of the
Protected Audience auction in which the auction waited for worklet
creation. To correctly capture this phase, we instrumented the
AuctionWorkletManager to report when the processes backing the worklets
became available. However, because of an incorrect use of
base::Unretained, the callback indicating that the process was available
could be called after the AuctionWorkletManager had already been cleaned
up following the end of the auction. This change replaces the
base::Unretained with a weak pointer provided by the existing
weak_ptr_factory_.

Bug: 348443991
Fixed: 357555097
Change-Id: Ic8b7d296b42cdeffd0412220c5b47b8e0f13dd9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5762309
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Orr Bernstein <orrb@google.com>
Cr-Commit-Position: refs/heads/main@{#1339837}