gpu: Remove context lost assert in FlushAndSubmit
https://crrev.com/c/5770817 handled null gr_context_ in FlushAndSubmit
under the assumption that this only happened due to context loss and
added a CHECK for that which started failing.

However, the root cause of the null gr_context_ seems to be that we
allow the client to create a OOP-R command buffer stub even if
SharedContextState wasn't initialized for Skia because
GrContextType::kNone is used when HW acceleration is disabled:

https://source.chromium.org/chromium/chromium/src/+/main:content/browser/gpu/gpu_data_manager_impl_private.cc;l=1376
https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/service/shared_context_state.cc;l=418

While a better fix is being worked on in https://crrev.com/c/5774422,
this CL removes the context lost assert to make it easier to merge to
stable channel.

Bug: 358349900, 355071122
Change-Id: I39c0cfe9e8b1fe7b3820e38083aee4fbe0ce4d9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5775963
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339862}