[Extensions][Reland] Add icons to MV2 deprecation disabled dialog
Chrome has announced that MV2 extensions will gradually no longer be
supported in the browser. None of these changes will take effect on any
Chrome channel before June, 2024. See more information at
https://developer.chrome.com/blog/resuming-the-transition-to-mv3.

The second stage of the deprecation shows a dialog listing the
disabled extensions. This CL adds the extension icon(s) to the dialog.
Since retrieving the icon is an async operation, we have to wait
for all icons to be retrieved before showing the dialog (and check
the extensions are still affected, since they could have been
updated externally).

Screenshots
- single extension:
https://drive.google.com/file/d/1J8K1nkUxBf80Hjpa9aWBvHiao008JIDC/view?usp=sharing
- multiple extensions:
https://drive.google.com/file/d/1mfiGDq8epQpFGrI12K3BzrWAA5SXOdXJ/view?usp=sharing

This is a reland of crrev.com/c/5741763, which introduced flaky tests
on ChromeOS due to a race condition with the 'informed restore dialog'
triggered on browser startup on Ash. While we continue working on a
fix for the flaky tests (tracked on crbug.com/358407359), this CL:
(1) Adds icons to the dialog
(2) Disabled the new tests on ChromeOS

Bug: 355218108, 358407359
Change-Id: I75ea0e613b27fb22995ec21e85a555a93c73c828
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5775618
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339867}