[views-ax] Migrate Role to new system in MediaProgressView
A recent CL https://chromium-review.googlesource.com/c/chromium/src/+/5759500
modifies the role exposed by MediaProgressView to kSlider, but it does
this by the old soon-to-be deprecated way of exposing a11y attributes.

This CL makes it so its using the new system, which caches the
attributes on the ViewAccessibility class.

This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: Ia3e3b50f31f1d3f4da7d01d41cb5f11c29ac201d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5777242
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Yiren Wang <yrw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339886}