Support null frame_entry cases in CreateNavigationRequestFromLoadParams.
It is sometimes possible to reach CreateNavigationRequestFromLoadParams
for subframes, such as when loading a PDF in a subframe (in which case
the PDF extension is loaded via NavigationController::LoadURL).

Per https://crbug.com/40467594, there are some cases when the
FrameNavigationEntry will be missing for subframes, which is causing
crashes in CreateNavigationRequestFromLoadParams. This CL adds support
for having a null frame_entry, until https://crbug.com/40467594 is
fixed.

Bug: 358084015, 40467594
Change-Id: I615316a812a846f691a0c13513d4fb5c3097c975
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5775515
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: Charlie Reis <creis@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339887}