Change Observer: Make the FileSystemAccessLockManager ref counted
The FileSystemAccessLockManager currently assumes that all locks will be
released before it is destroyed by the FileSystemAccessManager. This is
because all holders of locks are owned by FileSystemAccessManager who
orders its members such that the holders will be destroyed before the
FileSystemAccessLockManager.

But this is the wrong assumption since other sequences end up holding
locks. For example, FileSystemAccessFileHandleImpl::DidTakeSwapLock
passes it to DoFileSystemOperation by binding the lock to a callback.

The solution is to make FileSystemAccessLockManager ref counted. And
have the FileSystemAccessManagerImpl hold a scoped_refptr and each
RootLock hold a scoped_ref. That way the FileSystemAccessLockManager
stays alive as long as either the FileSystemAccessManagerImpl is alive
or if it has a lock.

Bug: 339898769
Change-Id: I80e49469d4e4439d21fe9fdaca1e00f7009e2f39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5775696
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Daseul Lee <dslee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339918}