p13n: cancel fullscreen preview when swiped left for SeaPen
When swiped left during fullscreen preview, personalization hub window
will route to the previous path.

In a previous CL, we added a wallpaper fullscreen element in
sea-pen-router to show preview for SeaPen images. The element will be
destroyed if the user swiped left to go from the SeaPen subpage to the
wallpaper subpage. Therefore, the popstate event cannot trigger the
action to turn fullscreen off.

The best (and easiest) way to fix this is to move the wallpaper
fullscreen element to personalization router. It also makes sense to
use one fullscreen element for all types of wallpapers.

BUG=b:355248369
TEST=Manually on DUT

Change-Id: I8764a7ca7a76c74ec28645bb165165ecbf8b93c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5778142
Reviewed-by: Jason Thai <jasontt@chromium.org>
Commit-Queue: Jerry Liu <pzliu@google.com>
Cr-Commit-Position: refs/heads/main@{#1339919}