Use delay timers to improve the experience for mouse clicking
If the user only wants to click on the progress view to seek to a new
position, we should not apply the visual changes for when the user is
dragging the progress view. Therefore we add a delay timer that delays
applying the changes after the press event.

We also add a delay timer for switching the progress view colors, since
when the user changes the media position, the media always pauses for a
short period to load the changes. In this way the progress view colors
will not change too often to cause glitch.

Bug: 355287478
Change-Id: I33b4f7290866fd78a373edc0cd258d4285b52db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5772113
Commit-Queue: Yiren Wang <yrw@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339921}