[Extensions WebIDL] Expand exceptions on unknown types
This CL adds exception raising during the schema compiler processing IDL
Type nodes if the processor doesn't recognize a type it encounters. Also
adds tests for these cases (which may need to be changed in the future
as more type support is added to the compiler).

To make this cleaner this CL also created a new custom Exception:
SchemaCompilerError and converts earlier Exceptions to use this more
specific exception class instead.

This is largely to help along with development, making it easy to see
when a Type node was not processed for some reason, but will also be
useful in the future for catching typos or people trying to use valid
WebIDL types we've not added support for yet.

Bug: 340297705
Change-Id: Id9eccc419e8707670fd02323df40484da79cb2a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5778152
Commit-Queue: Tim <tjudkins@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339946}