HttpStreamPool: Avoid creating Group/Job when QUIC/SPDY sessions exist
Before this CL, we created HttpStreamPool::{Group,Job} every time
RequestStream() and Preconnect() are called. If we already have
existing QUIC/SPDY sessions, we don't need to create them to handle
stream creations and preconnects because we can create an HttpStream
on top of them and there is no need to preconnect.

This CL add a thin implementation of HttpStreamRequest::Helper to
create HttpStreamRequest without creating a Group and a Job. This
should save both memory and CPU load a bit.

Existing tests should cover this change.

Bug: 346835898
Change-Id: Ib83117a01e684d02f1637ccd51a05d95259b7869
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5772633
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1339964}