drivefs_search: Inline retrying with offline query in `PerformSearch`
Currently, if the device is offline and a non-local-only DriveFS query
were to be performed in `PerformSearch`, it immediately calls
`OnSearchDriveFs` with `drive::FILE_ERROR_NO_CONNECTION`… which
immediately calls `PerformSearch` again, but with the query adjusted for
an offline search.

Inline this logic directly inside `PerformSearch` to prevent this
immediate re-call of `PerformSearch`. This reduces the number of returns
in `PerformSearch` down to one, making it easier to refactor.

Bug: b:357980197
Change-Id: I6a6a21005a3219b7609fd7c784fc60722f92a36a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5769155
Commit-Queue: Michael Cui <mlcui@google.com>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340010}