drivefs_search: Split search logic into separate query class
This is in preparation for allowing cancellation of DriveFS search
queries.

At the moment, all queries made by `DriveFsSearch::PerformSearch` are
never cancelled, as the `drivefs::mojom::SearchQuery` (now wrapped in a
`drivefs::DriveFsSearchQuery`) are "smuggled into" the `GetNextPage`
callback, ensuring that it will live until `GetNextPage` is called.

In a future CL (https://crrev.com/c/5769158), we will return this
`DriveFsSearchQuery` class so callers can cancel any search queries by
destroying the `DriveFsSearchQuery` if needed.

This query class currently takes in a pointer to the parent
`DriveFsSearch`, which is difficult to test - the tests would be
near-identical to the existing tests in `drivefs_search_unittest.cc`. A
future CL (https://crrev.com/c/5769157) will refactor this to take in a
pointer to a delegate instead, and introduce tests.

Bug: b:357980197
Change-Id: I6a6a21004a571f1053e631dd4c6a813c61ef3ce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5769156
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1340285}