Don't cache fallback monogram favicons
Prior to this CL, FaviconLoader was caching the fallback monogram
FaviconAttributes when it got one. This prevented requerying the
LargeIconService later on in the session (or rather, until a new cold
start).

This CL removes the caching of the fallback monogram version, and
requeries the LargeIconService each time the FaviconLoader is asked for
a favicon, until there is one returned by the LargeIconService.

Fixed: 358071092
Change-Id: Id4a644152094c5c74b698f0337ba5cca3794a671
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5776269
Auto-Submit: Louis Romero <lpromero@google.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340297}