[af] Fix awaiting calls calculation for Wait()
Include pending calls in the num_awaiting_total_calls counter as the
counter is only decremented when the call is done. This is to avoid
having the counter going down to 0 while there are still awaiting
calls that are expected.

For instance, if I Wait() for 3 calls while there is already 1
pending call and 0 total calls. The num_awaiting_total_calls
counter will go down to 0 when the first pending call is done,
stopping Wait() while there are still 2 calls that we want to wait
for.

Change-Id: Ib172cb497ef5dd8c290bdd911f0eb4a10d7ca85f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5777891
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1340312}