[iOS][QD] Animate partial tab group closure
This CL animates a tab group partially in cases where not all tabs in a
tab group are to be closed. In practise, we make the tab views inside
the tab group (either the snapshots or the favicon) disappear with the
animation.

Screencast: http://dr/file/d/12-qSwYI2BriM_KDIUx4vLfLvqZsRUaVI

Low-Coverage-Reason: HARD_TO_TEST Animations are hard to test.
Bug: 335387869, 354112735
Change-Id: Ie6d4b1a7fc047e28407fc89c0f88011199cd1e1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5766288
Reviewed-by: Zaina Al-Mashni <zalmashni@google.com>
Commit-Queue: Filipa Senra <fsenra@google.com>
Cr-Commit-Position: refs/heads/main@{#1340328}