[iOS] Add remove identity flow to AccountsTVC
This CL introduces the flow to remove an identity in the AccountsTVC.
To achieve that, this CL:
1- introduces AccountsMutator
2- introduces AccountsMediatorDelegate
3- Implements the account removal, later CL will add the confirmation
action sheet before actually removing the account.

The CL also fixes crash caused by not wiring the mediator to the
consumer in case kIdentityDiscAccountMenu is enabled.
(crbug.com/358091653 and crbug.com/358092088).

Bug: 349071402
Low-Coverage-Reason: TESTS_IN_SEPARATE_CL
Change-Id: Id667b2f251dad915c0bc5df3f8daff0ca36b5c2c
Fixed: 358091653, 358092088
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5769276
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1340335}