kiosk: Use explicit retry predicate in cryptohome checker
Currently cryptohome checker uses an optional to model errors to fit the
previous API required by RunUpToNTimes.

Use a 3-state enum to model the full result type of cryptohome checker
and use the new RunUpToNTimes that takes a `should_retry` predicate
function.

This renders the deprecated API unused, so we also remove it.

This is part of the effort to make kiosk_profile_loader unit testable.

Bug: 290172046
Change-Id: I75bbedfd5d86e323d17ecdbf05b17466bcedeb6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5771426
Commit-Queue: Edman Anjos <edman@chromium.org>
Reviewed-by: Irina Fedorova <irfedorova@google.com>
Cr-Commit-Position: refs/heads/main@{#1340341}