Chromium UI for integration with the MacOS SCContentSharingPicker
The integration with SCContentSharingPicker is built on the existing
DelegatedSourceListController mechanism.

The following Mac-specific adaptations to the UI have been made:
1. A button to open the MacOS picker is shown on the window and screen
   pane of the Chromium picker.
2. The reselect-button is not shown.
3. After a selection has been made, the dialog is automatically
   closed.

The actual implementation of the integration with
SCContentSharingPicker is added in a follow-up.
A skeleton of the DelegatedSourceListCapturer class is provided in
this change as an integration point.

Bug: 1494894
Change-Id: I4660088f02486f94689ec0d7422034bfc43df349
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5766075
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Tove Petersson <tovep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340351}