[iOS] Fix Parcel Tracking strings in What's New 128
The parcel tracking item for What's New M128 had some issues with its
Lottie asset strings, namely that the capitalization was different in
the .plist and .grdp than in the Lottie file, and there were extra
strings in the .plist and the .grdp that did not appear in the Lottie
file. This CL cleans all of that up in order to fix translations.

Bug: 356885355
Change-Id: Ida3202daebc5e866b80f8a683eeda2199514c676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5777484
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Cheick Cisse <cheickcisse@google.com>
Cr-Commit-Position: refs/heads/main@{#1340394}