Check that PlusAddressService is enabled before pre-allocating.
The PlusAddressService checks various conditions before it decides that
it is "enabled" - e.g., whether the user is signed in, whether the
current profile is really part of the Finch group, etc.
This CL allows injecting a check for the same properties into the
pre-allocator.

Bug: 324559503
Change-Id: I3bbfa145004f8903788c1fcc4b733297f202e801
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5776683
Reviewed-by: Kirubel Aklilu <kaklilu@chromium.org>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1340417}