Factor invalidation set construction into builder and visitor classes
Extract code in RuleFeatureSet that mutates RuleInvalidationData into
two new classes. Base class RuleInvalidationDataVisitor receives the
majority of the code. The eventual goal is to reuse as much of that
code as possible in a non-mutating form, in the case where we want to
reconstruct the relationships between style rules and invalidation
data for developer tooling. Derived class RuleInvalidationDataBuilder
receives code that makes up the entry points specific to invalidation
data mutation scenarios.

The intent of this CL is to bulk-move existing code to its new
locations with as few modifications as possible. Changes needed to
separate the builder and non-mutating tracer applications are left to
future CLs to make them more visible.

Bug: 337076014
Change-Id: Ida79202e88e63540f0d1b9daf14aaa6edbd4466d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5722834
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1340440}