android: orderfile: remove grouping by symbol name
Grouping by name is not used in the orderfile generator these days. It
is an anti-pattern because several symbols can appear at the same
offset.

Also remove the WarningCollector utility because it becomes unused. The
API is not ideal for current uses, but a similar utility may reappear
soon elsewhere.

Bug: None
Change-Id: Id2c409c910dbb9a90c81d1fa156132e7b58e1e00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5782836
Reviewed-by: Peter Wen <wnwen@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340443}