Update HTTPS-First Mode generated pref policy handling
With the addition of Balanced Mode, the enterprise policy handling is
slightly more complicated, and maps to two separate underlying prefs.
This updates the GeneratedHttpsFirstModePref logic to handle the
different cases when the Balanced Mode setting is available, and
reorganizes the logic to be clearer and better documented. This also
adds a unit test to exercise the different pref states that occur under
the different policy conditions (managed vs. recommended, and
"force_enabled", "force_balanced_enabled", or "disallowed").

Bug: 349860796
Change-Id: I188902cbfe3d0f5157299b22d8d5926bc76aca41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5774405
Commit-Queue: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1340459}